pyproject.toml
[tool.poetry]
name = "kaos"
version = "0.1.0"
description = ""
authors = ["npujol <naivy.luna@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.1.1"
django = "^5.0.4"
django-unfold = "^0.22.0"
django-import-export = "^3.3.8"
django-environ = "^0.11.2"
whitenoise = "^6.6.0"
gunicorn = "^21.2.0"
django-simple-history = "^3.5.0"
django-celery-beat = "^2.6.0"
django-modeltranslation = "^0.18.11"
django-debug-toolbar = "^4.3.0"
celery = {extras = ["redis"], version = "^5.3.6"}
ruff = "^0.3.7"
psycopg2 = "^2.9.9"
kombu = "^5.3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
lint.select = ["E", "F", "B", "UP", "DJ", "I", "PT"]
lint.ignore = ["E501", "UP007", "DJ001"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["A", "B", "C", "D", "E", "F", "UP", "DJ", "I", "PT"]
lint.unfixable = []
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]

# Same as Black.
line-length = 127

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.extend-per-file-ignores]
"shopaffe/settings_frontend.py" = ["F405"]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

#[tool.codespell]
## codespell supports pyproject.toml since version 2.2.2
## NOTE: the "tomli" package must be installed for this to work
## https://github.com/codespell-project/codespell#using-a-config-file
## NOTE: ignore words for codespell must be lowercase
#check-filenames = ""
#ignore-words-list = "word,another,something"
#skip = "htmlcov,.doctrees,*.pyc,*.class,*.ico,*.out,*.PNG,*.inv,*.png,*.jpg,*.dot"

[tool.coverage.run]
# excludes all files and directories matching these patterns
omit =[
    "*/migrations/*",
    "*/commands/*",
    "*/wsgi.py",
    "*/asgi.py",
    "**/tests/**"
]
# include is ignored if source (--cov=apps/feeds) is set